+++ src/components/intelligentMovement/MovementHandler.tsx
@@ -21,7 +21,7 @@ import { Animation } from '../../utils/pose/animations';
 import { ICoords } from '../../utils/types';
 import { getLargerDistance } from '../../utils/utils';
 import EyeController from '../eye/EyeController';
-import { analyseLight, naturalMovement } from '../eye/utils/EyeUtils';
+import { analyseLight, naturalMovement } from '../eye/utils/MovementUtils';
 
 interface IMovementProps {
     width: number;
